requirements.txt
streamlit
gradio
requests
beautifulsoup4
google-generativeai
python-dotenv
plotly
pandas
wordcloud
matplotlib
fpdf
